fix(web): 모임 찜/참여하기 상태 미반영 문제 해결
