+++ apps/web/src/_pages/moim-detail/index.ts
@@ -1,4 +1,5 @@
 export { CompactCard } from "./ui/compact-card";
 export { DescriptionSection } from "./ui/description-section";
 export { InformationContainer } from "./ui/information-container";
+export { MoimDetailClient } from "./ui/moim-detail-client";
 export { PersonnelContainer } from "./ui/personnel-container";
